requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
dataclasses-json==0.5.9
frozenlist==1.3.3
gensim==4.3.1
google-auth==2.21.0
google-auth-oauthlib==1.0.0
gspread==5.10.0
httplib2==0.22.0
idna==3.4
install==1.3.5
Janome==0.5.0
langchain==0.0.224
langchainplus-sdk==0.0.20
lxml==4.9.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mecab-python3==1.0.6
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.0
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.11
pyparsing==3.1.0
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.1
six==1.16.0
smart-open==6.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.18
tenacity==8.2.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==1.26.16
yarl==1.9.2
